requirements.txt
# Core dependencies for Squarest Notebook Simple
streamlit>=1.45.0
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.5.0
langchain>=0.3.3
langchain-openai>=0.2.3
langchain-anthropic>=0.2.3
langchain-google-genai>=2.0.1
langchain-community>=0.3.3
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.9.2
tiktoken>=0.8.0
humanize>=4.11.0
loguru>=0.7.2
httpx>=0.27.0
nest-asyncio>=1.6.0
